Templated the “Top” and “Bottom” of pages
Each page that has a header with the navigation menus and bottom with pictures, etc. is now templated out to a separate HTML file for easy editing that affects the whole site. There is a top.js and bottom.js for the loading function of each respective top.html and bottom.html templates.